Comienzo de manejo de permisos en back y cambio de Response Status

+++ paw-webapp/webapp/src/main/java/ar/edu/itba/paw/webapp/controller/UserController.java
@@ -50,7 +50,11 @@ public class UserController {
     @DELETE
     @Path("/users/{id}")
     public Response deleteUser(@PathParam("id") long id){
+        HogarUser hogarUser = (HogarUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(hogarUser.getUserID() != id){
+            return Response.status(Response.Status.FORBIDDEN).build();
+        }
         userService.deleteUser(id);
-        return Response.ok().build();
+        return Response.noContent().build();
     }
 }
